refactor: expose full public run API at top level
Re-export the remaining public run-management helpers from rush and
update public-facing docs/tests use top-level imports.
