+++ tests/test_auto3d.py
@@ -2,7 +2,7 @@ import itertools
 import sys
 
 from rush import auto3d
-from rush.runs import RunOpts
+from rush import RunOpts
 from tests._module_test_utils import assert_run_collects_and_caches
