+++ tests/test_exess_interaction_energy.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 from rush import FragmentRef, Topology, exess
-from rush.runs import RunOpts
+from rush import RunOpts
 from tests._module_test_utils import assert_run_collects_and_caches
