+++ tests/test_exess_interaction_energy_gadi.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import pytest
 
 from rush import FragmentRef, Topology, exess, fetch_run_info
-from rush.runs import RunOpts, RunSpec
+from rush import RunOpts, RunSpec
 from tests._module_test_utils import assert_run_collects_and_caches
