+++ tests/test_fold.py
@@ -1,5 +1,5 @@
 from rush import boltz, mmseqs2
-from rush.runs import RunOpts, RunSpec
+from rush import RunOpts, RunSpec
 from tests._module_test_utils import assert_run_collects_and_caches
