+++ tests/test_prepare_protein.py
@@ -2,7 +2,7 @@ from rush import TRC
 from pathlib import Path
 
 from rush import prepare
-from rush.runs import RunOpts
+from rush import RunOpts
 from tests._module_test_utils import assert_run_collects_and_caches
