+++ tests/test_prepare_protein_debump.py
@@ -1,7 +1,7 @@
 import math
 
 from rush import Element, from_pdb, prepare
-from rush.runs import RunOpts
+from rush import RunOpts
 from tests._module_test_utils import assert_run_collects_and_caches
